i_shape::float::simple

Trait SimplifyContour

pub trait SimplifyContour<P: FloatPointCompatible<T>, T: FloatNumber> {
    // Required method
    fn simplify_contour(&mut self, adapter: &FloatPointAdapter<P, T>) -> bool;
}

A trait that provides methods for simplifying complex geometrical structures.

Required Methods

fn simplify_contour(&mut self, adapter: &FloatPointAdapter<P, T>) -> bool

Simplifies the structure in-place if it is not already simple.

Returns

true if the structure was simplified successfully.

false if the structure was already simple and no modification was made.