i_shape::int::simple

Trait SimpleShapes

pub trait SimpleShapes {
    // Required methods
    fn is_simple(&self) -> bool;
    fn simplified(&self) -> IntShapes;
}

A trait for determining if a collection of shapes is simple, and for obtaining a simplified version of the entire collection.

Required Methods

fn is_simple(&self) -> bool

Checks if all shapes in the collection are simple.

Returns

true if all shapes are simple.

false if any shape in the collection is complex.

fn simplified(&self) -> IntShapes

Returns an optional simplified version of the collection.

Returns

IntShapes the simplified shapes.