i_shape::int::simple

Trait Simplify

pub trait Simplify {
    // Required method
    fn simplify_contour(&mut self) -> bool;
}

A trait that provides methods for simplifying complex geometrical structures.

Required Methods

fn simplify_contour(&mut self) -> bool

Simplifies the structure in-place if it is not already simple.

Returns

true if the structure was simplified successfully.

false if the structure was already simple and no modification was made.